Refactor addon structure
